+++ Makefile
@@ -268,7 +268,7 @@ STAGE_DIRS := src asm include linker tools assets
 OVERLAYS += checkps
 overlay_checkps_asset    := assets/checkps.bin
 overlay_checkps_gcc_srcs := src/overlays/checkps/unk1.c src/overlays/checkps/code3.c
-overlay_checkps_gnu_srcs := src/overlays/checkps/code4.c src/overlays/checkps/code5.c
+overlay_checkps_gnu_srcs := src/overlays/checkps/code4.c src/overlays/checkps/code5.c src/overlays/checkps/code6.c
 
 
 # ============================================================================
