+++ config/overlays/CHECKPS.BIN.yaml
@@ -46,6 +46,7 @@ segments:
       - [0xEA5, c, unk1]
       - [0x19B1, c, code5]
       - [0x1BC1, c, code3]
+      - [0x1DB5, c, code6]
       - [0x1F45, c, code2]
       - [0x27B9, rodata, table]
   - [0xA77D]
