+++ src/overlays/checkps/code4.c
@@ -2,6 +2,7 @@
 
 /**
  * decomp.me link (95%) https://decomp.me/scratch/EuGt8
+ * Matches 100% with GNU AS
  */
 void func_80050B04(void) {
     D_8005CFE8 = 1;
